refactor: Goal管理UIをGoalManagementViewに分離
